progressbar: fix mutex locking in {en|dis}able_progress_bar()
enable_progress_bar() and disable_progress_bar() used to be
only called by the info thread which was not terminated when
Squashfs exited prematurely on error or if ^C hit twice.  So we
did't need to have any pthread_cleanup_{push|pop} code.

Now, however, they are used by the ERROR_START() and ERROR_EXIT()
macros, which can be called by any thread, so we need to protect
the mutex lock against pthread cancel operations.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>