mksquashfs: fix delay in restoring filesystem if stuck in filesystem scanning
If when appending, control C is hit twice, Mksquashfs should respond
by restoring the filesystem and exiting.

Previously I discovered Mksquashfs might hang when generating the
filesystem for some time when control C was hit.  This
I discovered was due to the fact the Mksquashfs main thread might
not hit a pthread cancellation point for sometime.  I fixed this
by updating the restore thread to carefully terminate the
producer threads and flushing the queues that feed the Mksquashfs
main thread, so that it would quickly hit the pthread_cond_wait()
cancellation point on empty queues.

This works well, and avoids the messy need to insert an explicit
pthread_testcancel() within the filesystem generation code.

However, I have discovered Mksquashfs can still hang for sometime
if control C is hit when Mksquashfs is *scanning* the filesystem
prior to starting filesystem generation, if the source filesystem
is large.  This is again because the filesystem scanning code does
not execute any pthread cancellation points, and as it has not
yet started filesystem generation, it does not try and read any
data from producer queues.

The fix here is to exploit the fact that only at filesystem generation
time does the destination filesystem get written to, necessitating
it to be restored on termination.  Prior to this point we can safely exit
Mksquashfs without restoring the filesystem (because as yet it hasn't
been alterred).  So move the creation of the restore thread to
just before filesystem generation, with the result that prior to this
point Mksquashfs will exit without trying to restore the filesystem, and
we avoid trying to pthread cancel code which does not execute any pthread
cancellation points.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>