mksquashfs: sort_directory(), silence bogus cur NULL pointer dereference
Static analysis reports a cur NULL pointer dereference
at the cur->next=NULL line.

This is because static analysis is too stupid to realise we'll
always go around the loop at least once, assigning cur, iff
dir->count > 1, which is checked for.  In otherwords it does
not correlate dir->count > 1 with the fact dir->list will always
be non NULL.

Even though it is strictly unnecessary, it is easily to silence
static analysers by checking for dir->list == NULL as well as
dir->count > 1.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>