unsquashfs: remove_free_list(), silence bogus NULL pointer dereference
Static analysis reports a NULL pointer dereference in
remove_free_list().

This is because static analysis is too stupid to realise
if free_prev and free_next are both not NULL, then both
pointers are guaranteed to contain values.  It instead
determines a not checked for situation where one pointer can
contain a value and the other contains NULL, and flags this
as a potential NULL pointer dereference, even though this
situation cannot occur.

Even though it is strictly unnecessary, it is easily to silence
static analysers by changing the check to be free_prev == NULL ||
free_next == NULL.  The fact being it is sufficient for either
pointer to be NULL to know we have an empty list.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>