actions: add new absolute() test operator for symbolic links
Add new absolute() test operator that tests if a symlink has an
absolute path, which by definition means the symbolic link may
be broken (even if the absolute path does point into the filesystem
being squashed, because the resultant filesystem can be
mounted/unsquashed anywhere, it is unlikely the absolute path will still
still point to the right place).  If you know that an absolute symlink will
point to the right place then you don't need to use this function, and/or
these symlinks can be excluded by use of other test operators.

This is the easiest and most basic test if a symlink becomes dangling
test operator.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>