actions: add new contained_followlink() test operator for symbolic links
Test if a symlink appears to be within the source filesystem,
that is, it has a relative path, and the relative path does not
appear to backtrack outside the source filesystem using "..".

This test function evaluates the path for symlinks - that is it
follows any symlinks (and any symlinks that it may contain etc.), to
discover the canonicalised relative path.

This is expensive, if possible consider using the faster
contained() alternative which does not try to follow symlinks

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>