actions: add a prune action - allow fine tuning of excluding
This adds the prune action.  This action is designed to do fine
grained tuning of the in-core directory structure after the exclude,
move and pseudo actions have been performed.

Unlike exclude which can be most usefully used to exclude based on
the evaluation of the file in question in isolation, prune can be
used to do exclusion based on the overall state of the filesystem
once exclusion has been done, rather than just the single file.

This is for two reasons:

1. At exclude time the filesystem is being scanned and files are being
   evaluated and excluded.  As the filesystem has not been fully scanned,
   it is impossible to make decisons based on the state of other files.

   For example, delete directory "xyz" if it has 10 or fewer entries.
   At the time directory "xyz" is evaluated by the exclude action, the
   number of files that will be eventully be in it after scanning and
   exclusion is unknown.

   Or another example delete file "abc" if file "../../x/y/xxx" doesn't
   exist in the filesystem.  Again, as the filesystem has not yet
   been fully scanned this may not known when file "abc" is evaluted
   by the exclude action.

2. Many of the complex context aware test options not only
   are meaningless at exclusion time, but their complexity
   means they rely on the in-core directory to be present.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>