mksquashfs: move symlink reading from create_inode() to dir_scan1()
Up till now Mksquashfs did not read the contents of symlinks at
directory scanning, but left this task to the final filesystem
creation when the symlink inode in the output squashfs filesystem
was created.

Now that we're adding action test operations that operate on symlinks,
this creates a problem.  We want the values of the symlink when
evaluating the symlink for existence, recursive evaluation via
readlink() etc.  Up till now when implementing these tests I have
chosen to read the symlink from the source filesystem on demand and
then discard the value.  The over-arching reason for this is because
symlinks can reference other symlinks, and because these tests were
previously designed to be evaluated at exclude action time when the
directory structure has not been fully scanned, we need to deal with
symlinks that have not yet been scanned.  In other words moving
symlimk reading to directory scan time is no help when evaluating
synlinks at exclude time because there is no guarantee that any
referenced symlinks have been read.  You have no option but to fall back
to reading symlinks from the source filesystem at symlink test
evaluation time.

But evaluating symlinks by reading from the source filesystem is fraught
with difficulties, not least the fact that the existence of the
symlink in the source filesystem is no guarantee that the symlink exists
in the output filesystem.  Additionally there is never any guarantee
that the source filesystem hasn't changed whilst evaluating it rendering
any checks meaningless.

The fix to this has been to decide evaluating symlinks at exclude time
when the directory stucture has not been fully scanned causes
insurmountable problems.  The solution to this has been to introduce
a new prune action, which is evaluated on the fully scanned
directory structure.  This alleviates all the aforementioned
problems.

So, now that the prune action has been implemented, this checkin
moves reading of symlinks to the dir scanning phase, so that the
snapshotted values are available for the symlink test operations.

An additional minor improvement is that failure to read the symlink
for some reason is discovered at dir scanning time allowing Mksquashfs
to ignore the symlink.  Previously because reading the symlink was
performed at filesystem creation time, failure to read meant a
dummy empty symlink had to be created.  This is because at
filesystem creation all the metadata for the filesystem has been
computed and partially written, and this includes the count
of the number of inodes.

This is called a "minor improvement" because in practice this
situation never occurs because due to the nature of symlinks if Mksquashfs
could stat it at dir scanning time, then it is guaranteed to
be able to read it at filesystem creation time, unless the symlink
has been deleted in the meantime.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>