actions: rewrite exists() test operation
Rewrite exists() test operation to operate on the scanned in-core
directory structure.

This allows exists() to be rewritten, and to change it's behaviour to
be more straightforward.  It now tests to see if the symbolic link
exists in the output filesystem.  It also now follows any symbolic
links embedded in the pathname to ensure that the symbolic link
is fully dereferenceable to an existing entry, and that no
symlinks do not exist or are absolute.

In that way, it is believed the semantics of exists() are now more
obvious and are more intuitive.

The new implementation of exists() subsumes the previously separate
exists(), contained() and contained_followlink() test operations.

Previously, three separate test operations were needed, and they
were written in the way they were, due to limitations imposed due
to the fact they were operating on the source directory.  This is because
they were originally intended to be used with the exclude action where
the directory structure has not been fully scanned.

The introduction of the prune action allows these test operations to
be rewritten.  Which is, obviously, why the prune action was
introduced.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>