actions: fix uid() so it takes a name rather than just a number
The uid() test operation was added at the same time as all the other
inode attribute test operators, and all these operations were implemented
taking a [+-]number as an argument.

This is obviously wrong for UIDs.  UIDs should be able to be specified
as a name as well as a [+-]number.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>