action: expression logging should be using action->args
Fix display of the arguments in the action() statement.
Previously it used action->action->args which is the arg
count in the "generic" action statement, not the number of
actual arguments in the particular parsed action - i.e. the instance
rather than the definition.

For actions which take a fixed number of arguments there, obviously,
is no difference.  But for actions which take a variable number of
arguments the generic action count is "-2", and it is the
responsibility of an specific action parser helper to parse
the arguments.  In this case the discovered number of arguments
in the instance will be different to the -2 argument count in the
generic statement.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>